Added output of metadata for compression-related axes.
If an axis has "compress" attribute, then the dimensions listed
in this attributed must be present in the output file for the post-
processinbg to be able to decompress it. To make sure they are,
this modification parses the contents of "compress" attribute
and saves metadata for the axes found there.


Former-commit-id: 9a85719509fcbe6e461e5bdea0b41324a0e05a99