+Added verbosity flag for aof_set_coupler_flux
  Added an optional verbostity flag to aof_set_coupler_flux, to allow the
calling routine to control the extent to which this module dumps huge volumes of
output to stdout.  Also fixed the indents to help identify the extent of the
various subroutines and removed unused filler lines in atmos_ocean_fluxes.F90,
reducing the ovreall length of the file by about 30% and making it easier to see
what the code is actually doing.  All answers are bitwise identical.


Former-commit-id: 50cb35a91616b038fefff6b453dc12455297615e