Keep diag_manager from closing files that are not open
There are two cases when diag_manager attempted to close a file that
was not open:

1. If a file defined in the diag_table does not have any fields to
  write out
2. If a file name was defined more than once in the diag_manager

Both of these cases are now protected from attempting to close a file
that was previously opened.

The case in 1 is now protected by checking if the file is still open
(files(file)%file_unit is not -1).  Case 2 is now protected by
checking if a file with the same name is attempted to be registered
more than once.  If the file name is registered once, and ALL of the
other file options defined in the diag_table are identical, than only
a NOTE is printed.  If any of the file options are different, a FATAL
will be issued.

Unfortunately, since this is doing character compares, the checking for
multiple identical file names defined in a diag_table can take time
(if the diag_table is large), but as this is only done in the
initialization, the impact to the model run time should be minimal.


Former-commit-id: 64c0e617c62b803dc1d0f353a834e59ce60b9546