Address an issue with intel 15.0.3 compier that crashes in atmos_ocean_fluxes_init with -O2 in ocean/ice_null configuratin.
When compiled with -O2 using ifort 15.0.3 20150408 (Mac OSX), the model would crash
"malloc: pointer being freed was not allocated" at the end of atmos_ocean_fluxes_init, at
least in the configuration when ocean_null and ice_null are used (and of course there are
no ocean fluxes).

Some investigation shows that apparently part of the loop is evaluated ahead by the
optimized code (even though the body is not supposed to be executed), and the strings
constructed by concatenation are not properly allocated or deallocated by the compiler.


Former-commit-id: 70142de518683578636482ddaf3a1197d2a2d870