__APPLE__ replaced with autoconf test for gettid
On my Arch Linux distribution, the FMS build failed due to a pre-defined
gettid function, which refused to allow the build of a local gettid inside of
affinity.c.  (GCC 9.2, libc 2.30, Linux 5.3.5)

This patch adds a AC_CHECK_FUNCS test for gettid and redefines the local
version if it's unavailable.

HAVE_GETTID is defined for an automake build, but this patch may
potentially break any legacy MKMF builds for Darwin (OSX) users, unless
they manually define it in their templates.  (But perhaps that is the
preferred solution anyway?)


Former-commit-id: c68248150c9813b017a5c41326b6af408a35b668