Use SYS_gettid instead of __NR_gettid
The macro __NR_gettid is not defined on Mac OS.  Thus, switching to
SYS_gettid.  This _should_ be safe on Linux systems as: "The Linux kernel
header file defines macros `__NR_<name>`, but some programs expect the
traditional form `SYS_<name>`.  So in building libc we scan the kernel's
list and produce <bits/syscall.h> with macros for all the `SYS_` names.'


Former-commit-id: 71428a9cb54310d47dc03bcdbaaeb73ac7f21862