Remove __APPLE__ macro from Fortran and syscall on Mac
This commit removes all __APPLE__ macros from Fortran sources.  This
fix also removes a call to the deprecated `syscall` on Mac OS X.

__APPLE__ for the most part was used for the memory usage routines.
This fix confines all get memory calls to C, and changes
mpp_print_memuse_stats to be the only location where the C function is
called.  (Note, AFAIK the function to get the system memory is
portable to most *nix systems.)

The `syscall` function is deprecated in Mac OS.  The affinity.c source
was updated to not call `syscall` on Mac OS systems.


Former-commit-id: 53c842928b2c1b78918b89ea2f6d7ac0dc3373a1