Improve Mac OS support and remove __APPLE__ from Fortran sources (#249)
* Use SYS_gettid instead of __NR_gettid

The macro __NR_gettid is not defined on Mac OS.  Thus, switching to
SYS_gettid.  This _should_ be safe on Linux systems as: "The Linux kernel
header file defines macros `__NR_<name>`, but some programs expect the
traditional form `SYS_<name>`.  So in building libc we scan the kernel's
list and produce <bits/syscall.h> with macros for all the `SYS_` names.'

* Remove __APPLE__ macro from Fortran and syscall on Mac

This commit removes all __APPLE__ macros from Fortran sources.  This
fix also removes a call to the deprecated `syscall` on Mac OS X.

__APPLE__ for the most part was used for the memory usage routines.
This fix confines all get memory calls to C, and changes
mpp_print_memuse_stats to be the only location where the C function is
called.  (Note, AFAIK the function to get the system memory is
portable to most *nix systems.)

The `syscall` function is deprecated in Mac OS.  The affinity.c source
was updated to not call `syscall` on Mac OS systems.

* Add #ifdef around FC_FUNC macro

Added an ifdef around the define of FC_FUNC to allow non-Autoconf builds from failing.

* Add unit test for memuse.c

Added a simple test for the routine(s) in memuse.c

* Move memuse.c to mpp/mpp_memuse.c

The function in memuse.c is only called now by mpp.  Moving this file to the mpp directory, close to where it is used.

* Add mpp_memustil_mod tests

Added more mpp_memutils_mod tests.  Updated Makefiles for move of memuse.c to mpp/mpp_memuse.c

* Remove C version of memuse test

The C version of the test would like with libFMS using the C compiler.  This would cause issues with the libray name mangnling.  It can possibly be fixed, but memuse is called from Fortran in the library, and the Fortran test case should be sufficient.

* Add missing Linux case in mpp_memuse.c

Added the missing Linux preprocessor macros for mpp_memuse.c to use the correct method for reporting the max RSS memory

* Fix bad cpp macro name in #ifdef in mpp/mpp_memuse.c

Correct cpp macro name FC_FUNC in #ifdef

Adjusted comments to help future developers understand why things are done.

* Add documentation for new mpp test cases

This update only modifies the documentation in the new mpp test cases.

* Update documentation in mpp_mem_dump

Add better descrtiption of the mpp_mem_dump subroutine to the documentation.

* Add documentation to memustils/memutils.F90

Add documentation to memutils/memutils.F90

* Use Fortran standard for C/Fortran interoperability

This updated uses the Fortran standard to call C routines from Fortran.  Also includes more documentation in Doxygen for the updated files.

* Add license to memutils makefile.am

* Fix for data_override test

* Fix include file path in test_data_override.F90

Corrected the path to fms_platform.h in the #include statement

* Add missing -I to test data_override Makefile

Needed to find fms_affinity_mod module file.

* Have test_mpp_domains call fms_affinity_mod routines

* Add affinity to include for test_mpp_domains

* Fix data_override test after merge

A few routines were missing in the use/only statements after the merge with master.  This corrects that oversite

Update to a few fms2_io tests to remove warning messages.

* Add gx_f08_support.m4

Add new set of m4 macros checking for Fortran 2008 support in the Fortran compiler


Former-commit-id: aad1ebec3fb748eaedd45b293a28fb6bb973891f