mpp: BOZ literal assignment
Because BOZ literals (e.g. Z'00001000') are typeless, the standard does
not allow them to be assigned to variables, which do have prescibed
types.

While this has generally not been a problem in the past, GCC 10 now
forbids such assignments, which raises errors in the FMS build.

This patch resolves the issue by replacing assignments with an
`int(..,kind=...)` transfer.

Although this can be avoided with the `-fallow-invalid-boz` flag, the
following patch should avoid the need for any new flags.


Former-commit-id: 446e2f1709cc515fdafa2bf37e59486e6ea513e3